create components folder and move select image into it
